backend/requirements.txt
# RxMen Discovery Call Copilot - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# API Clients
anthropic==0.18.1              # Claude API
openai==1.12.0                 # OpenAI API (for embeddings)
pinecone>=3.0.0                # Pinecone vector database

# Data Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP & Async
httpx==0.26.0
aiofiles==23.2.1

# Data Processing
numpy==1.26.3
tiktoken==0.5.2                # Token counting for chunking

# PDF & Text Processing (already installed for extraction)
PyMuPDF>=1.24.0
PyPDF2==3.0.1
Pillow>=10.0.0

# Utilities
python-dateutil==2.8.2

# Development & Testing (optional, comment out for production)
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
